Updated to ignore .svn/ subdirectories or any sub-directories. Also fixed issue with lex-tag processing where additional user-provided attribute value pairs were getting dropped

+++ jcarafe-core/src/main/scala/org/mitre/jcarafe/util/SeparateDocs.scala
@@ -15,26 +15,28 @@ object SeparateDocs {
       val rr = ("<" + tg + " ").r
       val odirPath = args(1)
       dir.listFiles foreach { f =>
-        val sr = new java.io.FileInputStream(f)
-        var os = new java.io.OutputStreamWriter(new java.io.BufferedOutputStream(new java.io.FileOutputStream(new java.io.File(odirPath + f.getName + ".doc-" + cnt.toString))), "UTF-8")
-        val parser = new GenToker(sr)
-        var c = true
-        while (c) {
-          val t: Token = parser.getNextToken()
-          t.kind match {
-            case EOF => c = false
-            case TAGSTART =>
-              val ss = t.image
-              if (rr.findFirstIn(ss).isDefined) {
-                os.close
-                cnt += 1
-                os = new java.io.OutputStreamWriter(new java.io.BufferedOutputStream(new java.io.FileOutputStream(new java.io.File(odirPath + f.getName + ".doc-" + cnt.toString))), "UTF-8")
-              }
-            case TAGEND =>
-            case _ => os.write(t.image)
+        if (f.isFile) {
+          val sr = new java.io.FileInputStream(f)
+          var os = new java.io.OutputStreamWriter(new java.io.BufferedOutputStream(new java.io.FileOutputStream(new java.io.File(odirPath + f.getName + ".doc-" + cnt.toString))), "UTF-8")
+          val parser = new GenToker(sr)
+          var c = true
+          while (c) {
+            val t: Token = parser.getNextToken()
+            t.kind match {
+              case EOF => c = false
+              case TAGSTART =>
+                val ss = t.image
+                if (rr.findFirstIn(ss).isDefined) {
+                  os.close
+                  cnt += 1
+                  os = new java.io.OutputStreamWriter(new java.io.BufferedOutputStream(new java.io.FileOutputStream(new java.io.File(odirPath + f.getName + ".doc-" + cnt.toString))), "UTF-8")
+                }
+              case TAGEND =>
+              case _ => os.write(t.image)
+            }
           }
+          os.close
         }
-        os.close
       }
     } else println("usage: java -cp jcarafe...jar org.mitre.jcarafe.util.StripTags <input file> <output file>")
   }
